+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Require all HTTP clients to automatically decompress response bodies
 - Simplify Curl sentinel
+- Refactor test code
 
 ### Removed